vital_health_prediction/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
WTForms==3.1.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
pyserial==3.5
werkzeug==2.3.7
email-validator==2.0.0
bcrypt==4.0.1